requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
pydub==0.25.1
numpy==2.0.2
librosa==0.10.2.post1
tensorflow==2.18.0
vosk==0.3.45
requests==2.32.3
azure-storage-blob==12.10.0
gunicorn==20.1.0
